Merge back

+++ engines/access/amazon/amazon_player.cpp
@@ -4,10 +4,10 @@
  * are too numerous to list here. Please refer to the COPYRIGHT
  * file distributed with this source distribution.
  *
- * This program is free software; you can redistribute it and/or
- * modify it under the terms of the GNU General Public License
- * as published by the Free Software Foundation; either version 2
- * of the License, or (at your option) any later version.
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
  *
  * This program is distributed in the hope that it will be useful,
  * but WITHOUT ANY WARRANTY; without even the implied warranty of
@@ -15,8 +15,7 @@
  * GNU General Public License for more details.
  *
  * You should have received a copy of the GNU General Public License
- * along with this program; if not, write to the Free Software
- * Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA 02110-1301, USA.
+ * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  *
  */
 
@@ -48,7 +47,7 @@ void AmazonPlayer::load() {
 		_downDelta = -2;
 		_scrollConst = 2;
 
-		for (int i = 0; i < _vm->_playerDataCount; ++i) {
+		for (uint8 i = 0; i < _vm->_playerDataCount; ++i) {
 			_walkOffRight[i] = OVEROFFR[i];
 			_walkOffLeft[i] = OVEROFFL[i];
 			_walkOffUp[i] = OVEROFFU[i];
@@ -79,7 +78,7 @@ void AmazonPlayer::load() {
 		_diagDownWalkMax = 5;
 		_game->_guard->setPosition(Common::Point(56, 190));
 	} else {
-		for (int i = 0; i < _vm->_playerDataCount; ++i) {
+		for (uint8 i = 0; i < _vm->_playerDataCount; ++i) {
 			_walkOffRight[i] = SIDEOFFR[i];
 			_walkOffLeft[i] = SIDEOFFL[i];
 			_walkOffUp[i] = SIDEOFFU[i];
