+++ engines/ags/lib/allegro/file.cpp
@@ -4,10 +4,10 @@
  * are too numerous to list here. Please refer to the COPYRIGHT
  * file distributed with this source distribution.
  *
- * This program is free software; you can redistribute it and/or
- * modify it under the terms of the GNU General Public License
- * as published by the Free Software Foundation; either version 2
- * of the License, or (at your option) any later version.
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
  *
  * This program is distributed in the hope that it will be useful,
  * but WITHOUT ANY WARRANTY; without even the implied warranty of
@@ -15,8 +15,7 @@
  * GNU General Public License for more details.
  *
  * You should have received a copy of the GNU General Public License
- * along with this program; if not, write to the Free Software
- * Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA 02110-1301, USA.
+ * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  *
  */
 
@@ -40,7 +39,7 @@ int PACKFILE::pack_igetw() {
 	return pack_fread(buf, 2) == 2 ? READ_LE_UINT16(buf) : 0;
 }
 
-long PACKFILE::pack_igetl() {
+int32_t PACKFILE::pack_igetl() {
 	byte buf[4];
 	return pack_fread(buf, 4) == 4 ? READ_LE_UINT32(buf) : 0;
 }
@@ -52,7 +51,7 @@ int PACKFILE::pack_iputw(int w) {
 	return 0;
 }
 
-long PACKFILE::pack_iputl(long l) {
+int32_t PACKFILE::pack_iputl(int32_t l) {
 	byte buf[4];
 	WRITE_LE_UINT32(buf, l);
 	pack_fwrite(buf, 4);
@@ -64,7 +63,7 @@ int PACKFILE::pack_mgetw() {
 	return pack_fread(buf, 2) == 2 ? READ_BE_UINT16(buf) : 0;
 }
 
-long PACKFILE::pack_mgetl() {
+int32_t PACKFILE::pack_mgetl() {
 	byte buf[4];
 	return pack_fread(buf, 4) == 4 ? READ_BE_UINT32(buf) : 0;
 }
@@ -76,7 +75,7 @@ int PACKFILE::pack_mputw(int w) {
 	return 0;
 }
 
-long PACKFILE::pack_mputl(long l) {
+int32_t PACKFILE::pack_mputl(int32_t l) {
 	byte buf[4];
 	WRITE_BE_UINT16(buf, 4);
 	pack_fwrite(buf, 4);
@@ -207,7 +206,7 @@ int pack_igetw(PACKFILE *f) {
 	error("TODO: xxx");
 }
 
-long pack_igetl(PACKFILE *f) {
+int32_t pack_igetl(PACKFILE *f) {
 	return f->pack_igetl();
 }
 
@@ -215,7 +214,7 @@ int pack_iputw(int w, PACKFILE *f) {
 	return f->pack_iputw(w);
 }
 
-long pack_iputl(long l, PACKFILE *f) {
+int32_t pack_iputl(int32_t l, PACKFILE *f) {
 	return f->pack_iputl(l);
 }
 
@@ -223,7 +222,7 @@ int pack_mgetw(PACKFILE *f) {
 	return f->pack_mgetw();
 }
 
-long pack_mgetl(PACKFILE *f) {
+int32_t pack_mgetl(PACKFILE *f) {
 	return f->pack_mgetl();
 }
 
@@ -231,7 +230,7 @@ int pack_mputw(int w, PACKFILE *f) {
 	return f->pack_mputw(w);
 }
 
-long pack_mputl(long l, PACKFILE *f) {
+int32_t pack_mputl(int32_t l, PACKFILE *f) {
 	return f->pack_mputl(l);
 }
