+++ engines/ags/tests/test_inifile.cpp
@@ -4,10 +4,10 @@
  * are too numerous to list here. Please refer to the COPYRIGHT
  * file distributed with this source distribution.
  *
- * This program is free software; you can redistribute it and/or
- * modify it under the terms of the GNU General Public License
- * as published by the Free Software Foundation; either version 2
- * of the License, or (at your option) any later version.
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
  *
  * This program is distributed in the hope that it will be useful,
  * but WITHOUT ANY WARRANTY; without even the implied warranty of
@@ -15,17 +15,18 @@
  * GNU General Public License for more details.
  *
  * You should have received a copy of the GNU General Public License
- * along with this program; if not, write to the Free Software
- * Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA 02110-1301, USA.
+ * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  *
  */
 
 #include "ags/shared/core/platform.h"
-#include "ags/lib/std/algorithm.h"
-#include "ags/shared/debugging/assert.h"
+#include "common/std/algorithm.h"
+//#include "ags/shared/debugging/assert.h"
+// File not present??
+#include "common/scummsys.h"
 #include "ags/shared/util/file.h"
 #include "ags/shared/util/ini_util.h"
-#include "ags/shared/util/inifile.h"
+#include "ags/shared/util/ini_file.h"
 #include "ags/shared/util/stream.h"
 
 namespace AGS3 {
@@ -116,20 +117,20 @@ void Test_IniFile() {
 	delete fs;
 
 	// there are explicit sections and 1 implicit global one
-	const int section_count = 5;
+	const size_t section_count = 5;
 	// Test reading from the custom ini file
 	{
 		assert(ini.GetSectionCount() == section_count);
 		IniFile::ConstSectionIterator sec = ini.CBegin();
 
-		assert(sec->GetItemCount() == 1);
+		assert(sec->GetItemCount() == 1u);
 		IniFile::ConstItemIterator item = sec->CBegin();
 		assert(item->GetKey() == "global_item");
 		assert(item->GetValue() == "global_value");
 
 		++sec;
 		assert(sec->GetName() == "section1");
-		assert(sec->GetItemCount() == 5);
+		assert(sec->GetItemCount() == 5u);
 		item = sec->CBegin();
 		assert(item->GetKey() == "item1");
 		assert(item->GetValue() == "");
@@ -147,7 +148,7 @@ void Test_IniFile() {
 
 		++sec;
 		assert(sec->GetName() == "this_section_should_be_deleted");
-		assert(sec->GetItemCount() == 3);
+		assert(sec->GetItemCount() == 3u);
 		item = sec->CBegin();
 		assert(item->GetKey() == "item1");
 		assert(item->GetValue() == "value1");
@@ -159,7 +160,7 @@ void Test_IniFile() {
 
 		++sec;
 		assert(sec->GetName() == "section3");
-		assert(sec->GetItemCount() == 2);
+		assert(sec->GetItemCount() == 2u);
 		item = sec->CBegin();
 		assert(item->GetKey() == "item_to_be_deleted");
 		assert(item->GetValue() == "value");
@@ -169,7 +170,7 @@ void Test_IniFile() {
 
 		++sec;
 		assert(sec->GetName() == "section4");
-		assert(sec->GetItemCount() == 1);
+		assert(sec->GetItemCount() == 1u);
 		item = sec->CBegin();
 		assert(item->GetKey() == "item1");
 		assert(item->GetValue() == "value");
@@ -235,18 +236,18 @@ void Test_IniFile() {
 		ConfigTree tree;
 		IniUtil::Read("test.ini", tree);
 
-		assert(tree.size() == 5);
+		assert(tree.size() == 5u);
 		assert(tree.find("") != tree.end()); // global section
 		assert(tree.find("section1") != tree.end());
 		assert(tree.find("section3") != tree.end());
 		assert(tree.find("section4") != tree.end());
 		assert(tree.find("section5") != tree.end());
 		StringOrderMap &sub_tree = tree[""];
-		assert(sub_tree.size() == 1);
+		assert(sub_tree.size() == 1u);
 		assert(sub_tree.find("global_item") != sub_tree.end());
 		assert(sub_tree["global_item"] == "global_value");
 		sub_tree = tree["section1"];
-		assert(sub_tree.size() == 4);
+		assert(sub_tree.size() == 4u);
 		assert(sub_tree.find("item1") != sub_tree.end());
 		assert(sub_tree.find("item2") != sub_tree.end());
 		assert(sub_tree.find("item3") != sub_tree.end());
@@ -256,11 +257,11 @@ void Test_IniFile() {
 		assert(sub_tree["item3"] == "value3");
 		assert(sub_tree["new_item"] == "new_value");
 		sub_tree = tree["section3"];
-		assert(sub_tree.size() == 1);
+		assert(sub_tree.size() == 1u);
 		assert(sub_tree.find("item_to_be_kept") != sub_tree.end());
 		assert(sub_tree["item_to_be_kept"] == "another value");
 		sub_tree = tree["section4"];
-		assert(sub_tree.size() == 3);
+		assert(sub_tree.size() == 3u);
 		assert(sub_tree.find("new_item1") != sub_tree.end());
 		assert(sub_tree.find("item1") != sub_tree.end());
 		assert(sub_tree.find("new_item2") != sub_tree.end());
@@ -268,7 +269,7 @@ void Test_IniFile() {
 		assert(sub_tree["item1"] == "value");
 		assert(sub_tree["new_item2"] == "new_value2");
 		sub_tree = tree["section5"];
-		assert(sub_tree.size() == 3);
+		assert(sub_tree.size() == 3u);
 		assert(sub_tree.find("item5_1") != sub_tree.end());
 		assert(sub_tree.find("item5_2") != sub_tree.end());
 		assert(sub_tree.find("item5_3") != sub_tree.end());
